Moved Groups management to Group engine
