--- app/serializers/group_serializer.rb
@@ -1,8 +1,0 @@
-class GroupSerializer < BaseSerializer
-
-  schema do
-    type 'group'
-
-    map_properties :id, :name, :created_at, :updated_at
-  end
-end
