--- app/serializers/groups_serializer.rb
@@ -1,8 +1,0 @@
-class GroupsSerializer < BaseSerializer
-
-  schema do
-    type 'groups'
-
-    collection :groups, item, GroupSerializer
-  end
-end
