--- app/serializers/memberships_serializer.rb
@@ -1,8 +1,0 @@
-class MembershipsSerializer < BaseSerializer
-
-  schema do
-    type 'memberships'
-
-    collection :memberships, item, MembershipSerializer
-  end
-end
