+++ config/routes.rb
@@ -2,6 +2,7 @@ Katuma::Application.routes.draw do
 
   use_doorkeeper
 
-  mount Landing::Engine,    at: '/'
-  mount Account::Engine,    at: '/'
+  mount Landing::Engine, at: '/'
+  mount Account::Engine, at: '/'
+  mount Group::Engine,   at: '/'
 end
