+++ engines/group/README.rdoc
@@ -1,0 +1,3 @@
+= Group
+
+This project rocks and uses MIT-LICENSE.
