+++ engines/group/app/controllers/group/application_controller.rb
@@ -1,0 +1,6 @@
+module Group
+  class ApplicationController < ActionController::Base
+
+    include Shared::Controller::Authentication
+  end
+end
