+++ engines/group/app/helpers/group/application_helper.rb
@@ -1,0 +1,4 @@
+module Group
+  module ApplicationHelper
+  end
+end
