+++ engines/group/app/models/group/user.rb
@@ -1,0 +1,13 @@
+module Group
+  class User < ActiveRecord::Base
+
+    self.table_name = :users
+
+    include Shared::Model::ReadOnly
+
+    has_many :memberships,
+      dependent: :destroy
+    has_many :groups,
+      through: :memberships
+  end
+end
