+++ engines/group/app/policies/group/membership_policy.rb
