+++ engines/group/app/serializers/group/base_serializer.rb
@@ -1,0 +1,5 @@
+module Group
+  class BaseSerializer < Oat::Serializer
+    adapter FlatAPI
+  end
+end
