+++ engines/group/app/serializers/group/group_serializer.rb
@@ -1,0 +1,10 @@
+module Group
+  class GroupSerializer < BaseSerializer
+
+    schema do
+      type 'group'
+
+      map_properties :id, :name, :created_at, :updated_at
+    end
+  end
+end
