+++ engines/group/app/serializers/group/groups_serializer.rb
@@ -1,0 +1,10 @@
+module Group
+  class GroupsSerializer < BaseSerializer
+
+    schema do
+      type 'groups'
+
+      collection :groups, item, GroupSerializer
+    end
+  end
+end
