+++ engines/group/app/serializers/group/membership_serializer.rb
@@ -1,0 +1,10 @@
+module Group
+  class MembershipSerializer < BaseSerializer
+
+    schema do
+      type 'membership'
+
+      map_properties :id, :group_id, :user_id, :role, :created_at, :updated_at
+    end
+  end
+end
