+++ engines/group/app/serializers/group/memberships_serializer.rb
@@ -1,0 +1,10 @@
+module Group
+  class MembershipsSerializer < BaseSerializer
+
+    schema do
+      type 'memberships'
+
+      collection :memberships, item, MembershipSerializer
+    end
+  end
+end
