+++ engines/group/config/routes.rb
@@ -1,0 +1,9 @@
+Group::Engine.routes.draw do
+
+  namespace :api do
+    namespace :v1 do
+      resources :memberships, except: [:new, :edit]
+      resources :groups, except: [:new, :edit]
+    end
+  end
+end
