+++ engines/group/db/migrate/20150328134514_create_groups.rb
@@ -1,0 +1,9 @@
+class CreateGroups < ActiveRecord::Migration
+  def change
+    create_table :groups do |t|
+     t.string :name, null: false
+
+     t.timestamps
+    end
+  end
+end
