+++ engines/group/db/migrate/20150328134731_create_memberships.rb
@@ -1,0 +1,13 @@
+class CreateMemberships < ActiveRecord::Migration
+  def change
+    create_table :memberships do |t|
+      t.integer :user_id, null: false
+      t.integer :group_id, null: false
+      t.integer :role, null: false
+
+      t.timestamps
+    end
+
+    add_index :memberships, [ :user_id, :group_id ], unique: true
+  end
+end
