+++ engines/group/lib/group.rb
@@ -1,0 +1,4 @@
+require "group/engine"
+
+module Group
+end
