+++ engines/group/lib/group/version.rb
@@ -1,0 +1,3 @@
+module Group
+  VERSION = "0.0.1"
+end
