+++ engines/group/lib/tasks/group_tasks.rake
@@ -1,0 +1,4 @@
+# desc "Explaining what the task does"
+# task :group do
+#   # Task goes here
+# end
